criacao de classe para erros e adicionando validacoes na criacao de novos grupos

+++ src/storage/group/groupCreate.ts
@@ -1,11 +1,20 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
+import { AppError } from "@utils/AppError";
+
 import { GROUP_COLLECTION } from "@storage/config";
 import { groupsGetAll } from "@storage/group/groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup]);
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
